chore: moved lastUpdatedRef state from Editor to the subscription of relsoverStore
