+++ frontend/src/_stores/resolverStore.js
@@ -1,10 +1,15 @@
 import { create } from 'zustand';
+import { subscribeWithSelector } from 'zustand/middleware';
 import { createReferencesLookup, findAllEntityReferences, findEntityId } from './utils';
 import { createJavaScriptSuggestions } from '../Editor/CodeEditor/utils';
 import { v4 as uuid } from 'uuid';
 import _ from 'lodash';
 import { dfs, removeAppSuggestions } from './handleReferenceTransactions';
 import { deepClone } from '@/_helpers/utilities/utils.helpers';
+
+import { findComponentsWithReferences } from '@/_helpers/editorHelpers';
+
+import { flushComponentsToRender, useEditorStore } from '@/_stores/editorStore';
 
 class ReferencesBiMap {
   constructor() {
@@ -65,7 +70,7 @@ const initialState = {
 };
 
 export const useResolveStore = create(
-  (set, get) => ({
+  subscribeWithSelector((set, get) => ({
     ...initialState,
     actions: {
       updateStoreState: (state) => {
@@ -332,8 +337,67 @@ export const useResolveStore = create(
         });
       },
     },
-  }),
+  })),
   { name: 'Resolver Store' }
 );
+
+// Subscribed only to lastUpdatedRefs and compute the components that needs to be re-rendered
+useResolveStore.subscribe(
+  (state) => state.lastUpdatedRefs,
+  (lastUpdatedRefs) => {
+    if (lastUpdatedRefs.length > 0) {
+      const currentComponents =
+        useEditorStore.getState().appDefinition?.pages?.[useEditorStore.getState().currentPageId]?.components || {};
+
+      const directRenders = lastUpdatedRefs
+        .map((ref) => ref.includes('rerender') && ref.split(' ')[1])
+        .filter((item) => item !== false);
+
+      const toUpdateRefs = lastUpdatedRefs.filter((ref) => !ref.includes('rerender'));
+
+      const componentIdsWithReferences = findComponentsWithReferences(currentComponents, toUpdateRefs);
+
+      if (directRenders.length > 0) {
+        componentIdsWithReferences.push(...directRenders);
+      }
+
+      if (componentIdsWithReferences.length > 0) {
+        batchUpdateComponents(componentIdsWithReferences);
+      }
+    }
+  }
+);
+
+/**
+ ** Async updates components in batches to optimize and processing efficiency.
+ * This function iterates over an array of component IDs, updating them in fixed-size batches,
+ * and introduces a delay after each batch to allow the UI thread to manage other tasks, such as rendering updates.
+ * After all batches are processed, it flushes the updates to clear any flags or temporary states indicating pending updates,
+ * ensuring the system is ready for the next cycle of updates.
+ *
+ * @param {Array} componentIds An array of component IDs that need updates.
+ * @returns {Promise<void>} A promise that resolves once all batches have been processed and flushed.
+ */
+
+async function batchUpdateComponents(componentIds) {
+  if (componentIds.length === 0) return;
+
+  let updatedComponentIds = [];
+
+  for (let i = 0; i < componentIds.length; i += 10) {
+    const batch = componentIds.slice(i, i + 10);
+    batch.forEach((id) => {
+      updatedComponentIds.push(id);
+    });
+
+    useEditorStore.getState().actions.updateComponentsNeedsUpdateOnNextRender(batch);
+    // Delay to allow UI to process
+    await new Promise((resolve) => setTimeout(resolve, 0));
+  }
+
+  // Flush only updated components
+
+  flushComponentsToRender(updatedComponentIds);
+}
 
 export const useResolverStoreActions = () => useResolveStore.getState().actions;
